Move PathUtils and StreamUtils to internal package
