+++ src/main/java/ch/jalu/configme/internal/PathUtils.java
@@ -1,4 +1,4 @@
-package ch.jalu.configme.utils;
+package ch.jalu.configme.internal;
 
 import org.jetbrains.annotations.NotNull;
