+++ src/test/java/ch/jalu/configme/internal/PathUtilsTest.java
@@ -1,4 +1,4 @@
-package ch.jalu.configme.utils;
+package ch.jalu.configme.internal;
 
 import org.junit.jupiter.api.Test;
